Show remaining letters.  Validate input words.
Show the total number of tiles unseen by me (either on the
opponent's tray or not yet drawn).  The opponent can't have more
than 7 tiles on their tray, and if there are any undrawn then the
opponent must have exactly 7 on their tray.

Validate all of the words already on the board.